[0.3.0] Massive rewrite to support object dataList
So I made the fault of leaving my copy of this repository on google
drive which in turn did some magic and nuked my whole repo.
Which is why this commit has all the files from the repo in a single
commit. I hate myself right now and I am sorry for this shoddy commit.

Also, fuck you drive. T_T